Adapt Smithy output to distinguish between single and list fields
Lucene treats all fields as lists, so Elasticsearch/ECS's schema IDL
does not distinguish between fields that should be lists and those
that should not. This distinction matters for everything not based
on Lucene indices.

Until https://github.com/elastic/ecs/pull/661 lands in an official
ECS release, the model builder will identify which fields should be
rendered as Smithy lists based on an internal manifest.